+++ docs/manual-test.md
@@ -1,0 +1,15 @@
+# Manula Test Checklist
+
+---
+
+## MVP
+
+- [x] タスク名・内容・期限を入力してタスクを追加できる
+- [x] タスク名が空の場合、タスクは追加されない
+- [x] チェックボックスで完了・未完了を切り替えられる
+- [x] 完了済みタスクは見た目で区別できる
+- [x] 削除ボタンで対象タスクを削除できる
+- [x] タスクが0件でもエラーにならない
+- [x] リロード後もタスクが保持される
+- [x] 完了状態がリロード後も保持される
+- [x] 削除したタスクがリロード後に復活しない
